pyproject.toml
[project]
name = "nobunaga"
version = "0.1.0"
description = "Data pipeline + SPA for 信長之野望：真戰"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12",
    "httpx>=0.27",
    "openai>=1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "requests>=2.31",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
